agent: Fixes utmp/wtmp recording on 64-bit systems
This change fixes the incorrect formatting of utmp records on some
64-bit systems.

The change writes correct utmp records on:

  32-bit systems where session and time fields in utmp records are 32-bits
  64-bit systems where session and time fields in utmp records are 32-bits
  64-bit systems where session and time fields in utmp records are 64-bits

The only known case where a 64-bit system writes 64-bit session and time
fields is arm64/aarch64 architecture in, for example, Raspberry Pi
64-bit operating systems.

This change has been tested on:
  Debian 10.8.0 32-bit
  Debian 10.8.0 64-bit
  Fedora 33 64-bit
  Gentoo R-Pi 64-bit
  OpenSUSE Leap 15.2 64-bit
  R-Pi OS Latest 32-bit
  R-Pi OS Latest 64-bit
  Ubuntu 20.4 x86 LTS 64-bit
  Ubuntu 20.10 x86 64-bit
  Ubuntu 20.10 R-Pi 32-bit
  Ubuntu 20.10 R-Pi 64-bit

Fixes https://github.com/shellhub-io/shellhub/issues/765
